+++ ff-parser/src/test/java/uk/ac/ebi/uniprot/parser/ffwriter/line/cc/CCMassSpectromBuildTest.java
@@ -26,8 +26,8 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         List<String> evidences = new ArrayList<>();
         evidences.add(ev1);
         String note = "Monophosphorylated";
-        Double molWeight = 2189.4;
-        Double molWeightError = 0.0;
+        Float molWeight = 2189.4f;
+        Float molWeightError = 0.0f;
         MassSpectrometryCommentBuilder builder = buildComment(molWeight, molWeightError,
                 MassSpectrometryMethod.ELECTROSPRAY,
                 note, evidences, true);
@@ -52,8 +52,8 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         List<String> evidences = new ArrayList<>();
         evidences.add(ev1);
         String note = "Allele D, with 6 phosphate groups";
-        double molWeight = 22629;
-        Double molWeightError = 1.6;
+        Float molWeight = 22629f;
+        Float molWeightError = 1.6f;
         MassSpectrometryCommentBuilder builder = buildComment(molWeight, molWeightError,
                 MassSpectrometryMethod.ELECTROSPRAY,
                 note, evidences, true);
@@ -80,8 +80,8 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         List<String> evidences = new ArrayList<>();
         evidences.add(ev1);
         String note = "Allele D, with 6 phosphate groups";
-        double molWeight = 22629;
-        double molWeightError = 1.6;
+        Float molWeight = 22629f;
+        Float molWeightError = 1.6f;
         MassSpectrometryCommentBuilder builder = buildComment(molWeight, molWeightError,
                 MassSpectrometryMethod.ELECTROSPRAY,
                 note, evidences, false);
@@ -106,8 +106,8 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         List<String> evidences = new ArrayList<>();
         evidences.add(ev1);
         String note = "";
-        double molWeight = 3260;
-        Double molWeightError = null;
+        Float molWeight = 3260f;
+        Float molWeightError = null;
         MassSpectrometryCommentBuilder builder = buildComment(molWeight, molWeightError,
                 MassSpectrometryMethod.MALDI,
                 note, evidences, false);
@@ -131,8 +131,8 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         List<String> evidences = new ArrayList<>();
         evidences.add(ev1);
         String note = "Monophosphorylated";
-        Double molWeight = 871.3;
-        Double molWeightError = null;
+        Float molWeight = 871.3f;
+        Float molWeightError = null;
         MassSpectrometryCommentBuilder builder = buildComment(molWeight, molWeightError,
                 MassSpectrometryMethod.ELECTROSPRAY,
                 note, evidences, true);
@@ -147,7 +147,7 @@ public class CCMassSpectromBuildTest extends CCBuildTestAbstr {
         doTestString(ccLineString, comment);
     }
 
-    MassSpectrometryCommentBuilder buildComment(double molWeight, Double molWeightError,
+    MassSpectrometryCommentBuilder buildComment(float molWeight, Float molWeightError,
             MassSpectrometryMethod method,
             String note, List<String> evs, boolean setSources) {
         MassSpectrometryCommentBuilder builder =MassSpectrometryCommentBuilder.newInstance();
